Refactored tracer location; all tests passing
Removed Rack::Handler::WEBrick
