+++ lib/appmap/event.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 
 module AppMap
-  module Trace
-    MethodEventStruct =
-      Struct.new(:id, :event, :defined_class, :method_id, :path, :lineno, :static, :thread_id)
-
+  module Event
     @@id_counter = 0
 
     class << self
@@ -18,46 +15,8 @@ module AppMap
       end
     end
 
-    class Tracers
-      def initialize
-        @tracers = []
-      end
-
-      def empty?
-        @tracers.empty?
-      end
-
-      # TODO: remove functions argument
-      def trace(enable: true)
-        AppMap::Trace::Tracer.new.tap do |tracer|
-          @tracers << tracer
-          tracer.enable if enable
-        end
-      end 
-
-      def enabled?
-        @tracers.any?(&:enabled?)
-      end
-
-      def record_event(event, method: nil)
-        @tracers.each do |tracer|
-          tracer.record_event(event, method: method)
-        end
-      end
-
-      def delete(tracer)
-        return unless @tracers.member?(tracer)
-
-        @tracers.delete(tracer)
-        tracer.disable
-      end
-    end
-
-    class << self
-      def tracers
-        @tracers ||= Tracers.new
-      end
-    end
+    MethodEventStruct =
+    Struct.new(:id, :event, :defined_class, :method_id, :path, :lineno, :static, :thread_id)
 
     class MethodEvent < MethodEventStruct
       LIMIT = 100
@@ -68,7 +27,7 @@ module AppMap
 
           require 'appmap/util'
 
-          me.id = AppMap::Trace.next_id_counter
+          me.id = AppMap::Event.next_id_counter
           me.event = event_type
           me.defined_class = singleton ? AppMap::Util.descendant_class(method.owner).name : method.owner.name
           me.method_id = method.name.to_s
@@ -185,53 +144,6 @@ module AppMap
         super.tap do |h|
           h[:return_value] = return_value
         end
-      end
-    end
-
-    class Tracer
-      # Records the events which happen in a program.
-      def initialize
-        @events = []
-        @methods = Set.new
-        @enabled = false
-      end
-
-      def enable
-        @enabled = true
-      end
-
-      def enabled?
-        @enabled
-      end
-
-      # Private function. Use AppMap.tracers#delete.
-      def disable # :nodoc:
-        @enabled = false
-      end
-
-      # Record a program execution event.
-      #
-      # The event should be one of the MethodEvent subclasses.
-      def record_event(event, method: nil)
-        return unless @enabled
-
-        @events << event
-        @methods << method if method
-      end
-
-      # Gets a unique list of the methods that were invoked by the program.
-      def event_methods
-        @methods.to_a
-      end
-
-      # Whether there is an event available for processing.
-      def event?
-        !@events.empty?
-      end
-
-      # Gets the next available event, if any.
-      def next_event
-        @events.shift
       end
     end
   end
